Add contest_groups and prizes
Prizes are additional strings displayed in contest results based on
contestant's rank. They are intended to represent awards given in
official contests.

Since some official contests are held in stages, prizes are associated
not with contest directly, but with a new contest_group entity.

Currently contests to contest_groups many-to-one relationship is
modelled by a 'list of contests' field instead of a separate table.
This simplifies current code and is needed anyway for efficient
search of contest_groups. Still, a proper model may be desirable in the future.